nrd/rdv3: add chip addr offset to atu mappings
The System Info module is used to read system information such as Chip
ID and multichip mode. The ATU module needs to know the Chip ID to set
up the physical address of remote SCPs in case of a multichip scenario.
However, the System Info module is not available to the ATU during the
element init stage; it is only available during the module start stage.
Since ATU module needs chip id information during the element init stage
itself, read the SID register directly in the config data before passing
the elements to the ATU module.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I5835b0e9fafdcf7fb74b79517231d2ab10c6d8fa