neoverse-rd/rdv3: set L0GPTSZ to 16GB for all AP cores
For a RME enabled platform, supported L0GPT sizes are 1GB,
16GB, 64GB and 512GB. For RD-V3, the default L0GPTSZ field is set to 0,
which corresponds to each L0 table covering 1GB physical address range.
To address a 256TB address space (48bit PPS) and for 4KB page size (PGS),
the total size needed for the L0 table entries is 2MiB.
CCA security architect recommends keeping the L0 GPTs in the internal
SRAM, but the RD-V3 SRAM size is limited to 1MiB. To optimize the
number of L0 GPT entries required to address a 48bit PPS, set
L0GPTSZ to 4, which corresponds to each L0 GPT entries corresponding
to 16GB physical address space. The total of 128 KiB is required to
fit all the L0 GPT entries when L0GPTSZ is 4.

Setting it in the core manager's PE_STATIC_CONFIG register will
reflect in the GPCCR_EL3.L0GPTSZ system register of the AP.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Ibbbf0cd71f6eebfa7ec50df9219d157b3b470fb7