tc4: Add CCSM addresses
Add the base addresses of the CCSM block and the specific CCSMs for each
CPU type and add this base address to the ATU configuration to allow SCP
access. The CCSM ATU region size is set to 8MiB, as per specification.

This commit also adds CCSM addresses for DSU, CME and the various GPU
CCSM modules.

Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>