nrd/rdv3: platform module to bind with SCP-RSE SCMI channel
SCP platform module to bind with SCP-RSE SCMI channel. SCP uses this
channel to report shutdown and cold reboot requests to RSE.

The current implementation of RSE do not support SCMI. This patch can be
removed once SCMI support is enabled in RSE. In that scenario, RSE will
subscribe to the system power-down notifications, which will then be
handled directly by the SCP SCMI module.

Change-Id: Ib912ddb668c88ed036d0876c0824ae743ff06bdd
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>