interface/power_management: Add set power demand API
Add an API to set power demand for an entity in the system
via the generic fwk id.

This API should be implemented in components that care for the
power demand of an entity. And API should be called by components
that want to communicate the information of the power demand to
those components mentioned earlier.

Signed-off-by: Mohamed Omar Asaker <mohamed.omarasaker@arm.com>