mod/system_coordinator: Add initial module
This patch adds the system coordinator module to the power management
framework. The system coordinator is responsible for managing each
phase of power management.

Each phase is an API call to a module. The phase API and time are
configurable elements. Coordinator calls the next phase after the
phase timer alarm elapses. The order of the phases follows the element
order.

A cycle timer alarm periodically notifies the coordinator to call all
the phases again. The cycle time is configurable and must be more than
the sum of all phase time.

Signed-off-by: Wen Ping Teh <wenping.teh@arm.com>