mod_power_capping_hal: Use PID controller in Power Capping Hal.
Apply PID feedback on Power Cap requested by the agent as follows:
1) The set point is the requested cap by the agent.
2) The feedback is the average measured power.
3) The output is a power limit that is delivered to the power distributor.

The usage of the PID would allow the Power Cap not to be set directly
to the system although a margin would be there to allow bursts of performance
applied on the system for a short period of time.

The PID is called continuously on intervals inside the get_power_limit function
called from metrics_analyzer.

Signed-off-by: Mahmoud Elsabbagh <mahmoud.elsabbagh@arm.com>