mod_perf_controller: Rename perf controller blocks.
The renaming was carried out as perf controller use cases
grow larger, its components names need to be renamed to adapt
to the new use cases. The blocks core and cluster would be
renamed limiter and controller domain to avoid confusion and to
be more generic.

The controller domain in the system consists of various performance
limiters and applies the lowest performance limit across the domain.

Signed-off-by: Mahmoud Elsabbagh <mahmoud.elsabbagh@arm.com>