mod/power-distributor: Add tree traverse order table
Add the construction of the tree traverse order table.
The table shall be used during calculations and distribution of
limits, demands, and budgets.

Traverseing order will be level by level (breadth first walk).

The construction is based on detecting the root and adding it to
the traverse order table then looping through the table adding
children to the end.

Unit test added.

Signed-off-by: Mohamed Omar Asaker <mohamed.omarasaker@arm.com>