mod/ccsm: Fix bitwise vs logical OR issue and internal assert
A bitwise OR operator has been used instead of the logical one
inside an if statement in ccsm_mod_set_configuration().

Also an assert meant to check for pointers passed as parameters
to be non NULL failed to properly do that, allowing for a
potential NULL pointer dereference later on in
get_register_value_rate().

Signed-off-by: Francesco Rosati <francesco.rosati@arm.com>