mod_scmi: Fix the number of SCMI notifications sent to an agent
SCMI notification should be sent to the subscribers for a specific
operation on a specific resource. But, the current SCMI notification
has a defect, it sends notification for a specific operation on any
resource, resulting in sending same notification multiple times to
the subscribers.

The looping around element/resource list in SCMI Notification table
is redundant, hence omitted to fix the defect.

Signed-off-by: Md Golam Tanvir Zim <mdgolamtanvir.zim@arm.com>
Change-Id: I025a77393a9396f7723827eaf4e6ce3973dd1ac7