module/atu: fix ATU encode attributes macro
The AxNSE and AxPROT[1] attributes programmed for an ATU region are used
to determine the physical address space(PAS) of the transaction. Other
attributes such as AxCACHE[x] and AxPROT[x] must be set to 'passthrough'
so that the value at the input bus is transferred to the output bus as
is.

The default ATU_ENCODE_ATTRIBUTES macros incorrectly set the other bus
attributes to zero instead of passthrough. So, fix this by updating the
following macros:
- ATU_ENCODE_ATTRIBUTES_ROOT_PAS
- ATU_ENCODE_ATTRIBUTES_SECURE_PAS
- ATU_ENCODE_ATTRIBUTES_NON_SECURE_PAS
- ATU_ENCODE_ATTRIBUTES_REALM_PAS

Signed-off-by: Shriram <shriram.k@arm.com>
Change-Id: I310bd2315feba01c95999813d6d19bdce999f79d