module/atu: fix output bus attribute configuration
In the ATU manage mode, the output bus attributes for an ATU region is
set in the ATUROBA register without clearing the existing configuration.
In certain scenarios, this may lead to incorrect output bus attributes
being configured for an ATU region when an ATU region with different
output bus attributes had been programmed previously in that register.

So, fix the ATU region output bus attributes programming in the
atu_add_region() function.

Signed-off-by: Shriram <shriram.k@arm.com>
Change-Id: I7627068712aacbfc7715436118c47d8fcbdcae11