mod/scmi_system_power: Refactor SCMI System Power Policy function
The policy function in mod_scmi_system_power.c is intended to
be overridden by a platform specific implementation of the
function. However, the default policy function is able to send
notifications to a platform through scmi_sys_power_notify()
which is a static function. Moving its call to outside the
policy function makes it accessible to any platform specific
policy function that wishes to make use (or not) of it,
depending on the override function.

Signed-off-by: Mubashar Ahmad <mubashar.ahmad@arm.com>