arch/aarch64: Support nested exceptions
Currently, if a second exception handler is called immediately after
the first, the core gets stuck in irq_global as this is where the second
exception gets called from. Similarly if a second exception were to
pre-empt the first then execution would get stuck inside the first
exception handler.

To fix this, save and restore SPSR_EL2 and ELR_EL2 inside fiq_handler.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>