prod/fmu: Refactor FMU module to support multiple implementations
It is known that different FMUs exist with different features and
hardware interfaces, e.g. the GIC FMU and eventually the MHU FMU and NIC
FMUs.

Create an internal API struct (in fmu_common.h) and refactor the current
implementation into two - a System FMU implementation and an
SCP-firmware API which forwards calls to a specific implementation based
off the device's configuration.

Update all test cases to follow.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>