arch/aarch64: remove mpidr_el1 check at boot time
mpidr_el1 value was check at boot time to ensure that the firmware was executed
on the primary core.
This check prevents booting on Renesas/Salvator platform (cortex A53/A57)
because the SCP firmware is not executed on the main processor.
Moreover, this check is useless on other platforms supporting Aarch64 as they
use a single processor element.

Signed-off-by: Florian Sylvestre <fsylvestre@baylibre.com>